+++ NEWS.md
@@ -1,5 +1,12 @@
 # rnndescent (development version)
 
+## New features
+
+* New parameter: for `nnd_knn` and `rnnd_build`: `weight_by_degree`. If set
+to `TRUE`, then the candidate list in nearest neighbor descent is weighted in
+favor of low-degree items, which should make for a more diverse local join.
+There is a minor increase in computation but also a minor increase in accuracy.
+
 ## Bug fixes and minor improvements
 
 * Fix a factor of 2 error in the TS-SS metric. This does not affect the returned
@@ -53,9 +60,9 @@ simplified.
 ## New features
 
 * New function: `rnnd_knn`. Behaves a lot like `rnnd_build`, but *only* returns
-the knn graph with no index built. The index can be very large in size for high
-dimensional or large datasets, so this function is useful if you only care about
-the knn graph and won't ever want to query new data.
+the knn graph with no index built. The index can be very large in size for
+high dimensional or large datasets, so this function is useful if you only
+care about the knn graph and won't ever want to query new data.
 * New function: `neighbor_overlap`. Measures the overlap of two knn graphs via
 their shared indices. A similar function was used extensively in some vignettes
 so it may have sufficient utility to be useful to others.